added a lot of functionality
The code has changed quite a bit as I've added complication and moved values into arrays (easier to reference later). I'm learning as I go along, how fun!

Added a bunch of new functionality, a little menu at startup and 5 different modes of operation (in total). The modes are:
1 = MIDI Note ON+OFF (timed)
2 = MIDI Note Toggle (must hit switch again to turn off)
3 = MIDI CC ON+OFF (timed)
4 = MIDI CC Toggle (same as MIDI Note, but with CC)
5 = Program Change

The default mode is set by the RUNMODE_DEFAULT variable (currently set to 1). The default mode will be highlighted on startup and there is a choice to choose a different mode before the RUNMODE_TIMEOUT timer runs out.